chore: separe repository from usecase
