+++ pkg/domain/accounts/usecases.go
@@ -19,11 +19,3 @@ type UseCase interface {
 	GetByID(ctx context.Context, id string) (entities.Account, error)
 	GetAll(ctx context.Context) ([]entities.Account, error)
 }
-
-type Repository interface {
-	GetAll(ctx context.Context) ([]entities.Account, error)
-	GetBalanceByID(ctx context.Context, id string) (int, error)
-	GetByID(ctx context.Context, id string) (entities.Account, error)
-	CheckCPF(ctx context.Context, cpf string) error
-	Upsert(ctx context.Context, account entities.Account) error
-}
